Views e Controller Preenchidos

+++ database/migrations/2022_03_19_001006_create_opcoes_table.php
@@ -15,7 +15,7 @@ class CreateOpcoesTable extends Migration
     {
         Schema::create('opcoes', function (Blueprint $table) {
             $table->id();
-            $table->integer('opcao_id')->unsigned();
+            $table->bigInteger('opcao_id')->unsigned();
             $table->foreign('opcao_id')->references('id')->on('enquetes');
             $table->string('opcao');
             $table->integer('votos')->default(0);
